+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 ### Fixed
-
+- Fixed an issue in Static DatePicker (MUI) where the "Today" action did not correctly update or reflect the selected date when triggered from the year/month view.
 ### Changed
 
 ### Breaking changes